Feat : 유저 평가 결과 취합 조회

+++ backend/module-api/src/main/java/com/ssafy/api/request/EvaluationBase.java
@@ -12,5 +12,5 @@ import lombok.Setter;
 public class EvaluationBase {
     Long parentId;
     Long id;
-    Integer point;
+    Float point;
 }
